successfully handled navigate to another page when error message is user already exists
